requirements.txt
joblib==1.1.0
matplotlib==3.5.2
numpy==1.22.3
pandas==1.4.2
Pillow==9.1.0
streamlit==1.9.0
